Casting changes to fix compilation issues under MacOS X. From N Harman.
git-svn-id: http://svn.python.org/projects/stackless/Python-2.4.2/dev@44693 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Stackless/core/stackless_impl.h
@@ -38,7 +38,7 @@ PyAPI_DATA(int) slp_try_stackless;
 PyAPI_DATA(PyCStackObject *) slp_cstack_chain;
 
 PyAPI_FUNC(PyCStackObject *) slp_cstack_new(PyCStackObject **cst,
-					    int *stackref,
+					    intptr_t *stackref,
 					    PyTaskletObject *task);
 PyAPI_FUNC(size_t) slp_cstack_save(PyCStackObject *cstprev);
 PyAPI_FUNC(void) slp_cstack_restore(PyCStackObject *cst);
